Moved common tracepoint support into tracepoint.py
